refactor: 키워드 입력 컴포넌트 모듈화
